ensure Junos API client/transport libraries always load
All files in `libraries` are evaluated on all Chef runs. In the case of
this cookbook you may end up loading files not intended for your
platform. For example if you are using this cookbooks on an Arista
switch you won't have the `junos-ez-stdlib` gem installed we don't want
the Chef run blow up.