Create a separate *_name attribute for all resources
This should not be confused with `Chef::Resource#name` which most every
resource has. This is meant to hold what can be thought of as the unique
"id" that uniquely identifies the resource on the switch. In almost all
cases the `#TYPE_name` attribute value will match that of `#name`.

This follows the pattern of other resources in Chef including: package,
service, etc.